+++ NEWS.md
@@ -1,13 +1,22 @@
 # log history of echarty package development
 
-## echarty 0.1.1
+## v. 0.1.2
+
+- ec.data now with three format values
+- ec.fromJson has now "..." (additional arguments)
+- ecr.band has new parameter 'two' for alternative rendering
+- error bars added as ecr.ebars with support for grouped data
+- more examples in ec.examples
+- some code optimization
+
+## v. 0.1.1
 
 - added ec.plugjs - install unknown plugins, like JS maps
 - added ec.js2r - a JS to R translation assistant, a Shiny app
 - ec.init update: one-time install for all plugins except built-in leaflet/custom
 - GL merged with 3D as plugin '3D'
-- more plugins: liquidfill, world, wordcloud, gmodularity
+- more plugins: liquidfill, world, wordcloud, gmodular
 - ec.sband renamed to ec.band 
 
-## echarty 0.1.0
+## v. 0.1.0
 - initial release, approved by CRAN Feb 16, 2021